+++ packages/analytics-node/src/config.ts
@@ -1,13 +1,17 @@
 import { Config, NodeOptions, NodeConfig as INodeConfig } from '@amplitude/analytics-core';
-import { Http } from './transports/http';
+import { DEFAULT_REQUEST_TIMEOUT_MILLIS, Http } from './transports/http';
 
 export class NodeConfig extends Config implements INodeConfig {
+  requestTimeoutMillis: number;
+
   constructor(apiKey: string, options?: NodeOptions) {
+    const requestTimeoutMillis = options?.requestTimeoutMillis ?? DEFAULT_REQUEST_TIMEOUT_MILLIS;
     super({
-      transportProvider: new Http(),
+      transportProvider: new Http(requestTimeoutMillis),
       ...options,
       apiKey,
     });
+    this.requestTimeoutMillis = requestTimeoutMillis;
   }
 }