+++ packages/analytics-react-native/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.7.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-react-native@1.6.9...@amplitude/analytics-react-native@1.7.0) (2026-08-24)
+
+
+### Features
+
+* **analytics-react-native:** add diagnostics client ([#1939](https://github.com/amplitude/Amplitude-TypeScript/issues/1939)) ([8d150ce](https://github.com/amplitude/Amplitude-TypeScript/commit/8d150ce6611b628e9e10a59ea96da07ec71e9fcf))
+
+
+
+
+
+## [1.6.9](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-react-native@1.6.8...@amplitude/analytics-react-native@1.6.9) (2026-08-12)
+
+**Note:** Version bump only for package @amplitude/analytics-react-native
+
+
+
+
+
 ## [1.6.8](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-react-native@1.6.7...@amplitude/analytics-react-native@1.6.8) (2026-07-28)
 
 **Note:** Version bump only for package @amplitude/analytics-react-native